# docs/openapi.yml
openapi: 3.0.3
info:
  title: Tournament Platform API
  version: v1
  description: |
    API for the Tournament Platform, enabling user management, wallet operations, tournament participation, and administrative functions.
    Key features include secure authentication, real-time match updates, and role-based access control.
  contact:
    name: Support Team
    email: support@example-tournament.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:{port}/api/v1 # Define port in variables or replace directly
    description: Local development server
    variables:
      port:
        default: '3000'
        description: Port for the local server
  - url: https://staging.example-tournament.com/api/v1
    description: Staging server
  - url: https://api.example-tournament.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access Token for authentication. Example: 'Bearer {token}'"
    refreshTokenCookie: # For documenting the refresh token mechanism
      type: apiKey
      in: cookie
      name: jid # Matches JWT_REFRESH_COOKIE_NAME from .env.example
      description: "HttpOnly cookie containing the Refresh Token."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Validation Error"
        errors: # Optional, for detailed validation errors
          type: array
          items:
            type: string
          example: ["'email' is required"]
        stack: # Optional, for development environment
          type: string

    # Auth Schemas
    UserRegistrationRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          example: "player123"
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "Str0ngP@sswOrd!"
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "Str0ngP@sswOrd!"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        accessToken:
          type: string
          format: jwt
          description: "Short-lived Access Token (typically sent in response body for SPA, or handled by server for cookie-based)"
        user:
          $ref: '#/components/schemas/UserPublicProfile'
      # Note: Refresh token is typically set in an HttpOnly cookie and not part of the JSON response.
...
    UserPublicProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        username:
          type: string
          example: "player123"
        role:
          type: string
          enum: [User, Admin, DisputeModerator, FinanceManager]
          example: "User"

    # Wallet Schemas
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User registered successfully. Access token provided, refresh token set in cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie: # Documenting the refresh token cookie
              schema:
                type: string
                example: jid=yourRefreshToken; Path=/; HttpOnly; Secure; SameSite=Strict
        '400':
          description: Validation error (e.g., email already exists, weak password).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in and receive Access and Refresh Tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful. Access token provided, refresh token set in cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jid=yourRefreshToken; Path=/; HttpOnly; Secure; SameSite=Strict
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Get a new Access Token using a Refresh Token.
      description: Requires a valid Refresh Token sent via an HttpOnly cookie (`jid`).
      security:
        - refreshTokenCookie: [] # Indicates cookie is used
      responses:
        '200':
          description: New access token generated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    format: jwt
        '401':
          description: Invalid or expired refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Invalidate the Refresh Token for a secure logout.
      description: Requires authentication with an Access Token. Clears the refresh token cookie.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful. Refresh token cookie cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
          headers:
            Set-Cookie: # Shows cookie being cleared
              schema:
                type: string
                example: jid=; Path=/; HttpOnly; Expires=Thu, 01 Jan 1970 00:00:00 GMT
        '401':
          description: Unauthorized (no valid access token).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...